Support for custom groups
Adding support for custom groups to the AddLayerCheckbox method.  Custom group will be build if it is not found, will persist between event/editing mode and the checked status will persist across hard refreshes (will be reset upon browser close/open).  Child elements will now also check against the checked status of the parent when they are added and disable themselves if the parent group is not enabled.